major refactoring of concrete module
