+++ compiler/src/scope/concrete/mod.rs
@@ -1,0 +1,3 @@
+pub mod callable_registry;
+pub mod core;
+pub mod struct_registry;
